requirements.txt
flask==2.0.1
werkzeug==2.0.1
sqlalchemy==1.4.46
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.5
python-dotenv==1.0.0